refactor: custom hook에서 validate 함수 분리
