+++ hooks/src/lib/hooks/useCardCVC.ts
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import { validateCVC } from '../validation/validateCVC';
 
 const useCardCVC = () => {
   const [cardCVC, setCardCVC] = useState('');
@@ -6,24 +7,9 @@ const useCardCVC = () => {
   const [errorMessage, setErrorMessage] = useState('');
 
   const handleCVCValidate = (input: string) => {
-    if (!validateCVC(input)) return;
+    if (!validateCVC({ input, setIsValid, setErrorMessage })) return;
     setIsValid(true);
     setErrorMessage('');
-  };
-
-  const validateCVC = (input: string) => {
-    if (Number.isNaN(Number(input))) {
-      setIsValid(false);
-      setErrorMessage('숫자만 입력해 주세요.');
-      return false;
-    }
-    if (input.length !== 3) {
-      setIsValid(false);
-      setErrorMessage('3글자의 숫자를 입력해 주세요.');
-      return false;
-    }
-
-    return true;
   };
 
   return { cardCVC, setCardCVC, handleCVCValidate, isValid, errorMessage };
