+++ hooks/src/lib/hooks/useCardNumber.ts
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import { validateCardNumber } from '../validation/validateCardNumber';
 
 interface CardNumberInput {
   input1: string;
@@ -26,57 +27,10 @@ const useCardNumber = () => {
 
   const handleCardNumber = (numbers: CardNumberInput) => {
     setCardNumber(numbers);
-    validateCardNumber(numbers);
-  };
-
-  const validateCardNumber = (numbers: CardNumberInput) => {
-    let hasError = false;
-    let message = '';
-
-    const newIsValid = { ...isValid };
-
-    if (!isValidFirstCardNumber(numbers.input1)) {
-      newIsValid.input1 = false;
-      message = '첫번째 카드 번호는 4 또는 51~55 사이의 숫자여야 합니다.';
-      hasError = true;
-    } else {
-      newIsValid.input1 = true;
-    }
-
-    (Object.entries(numbers) as [keyof CardNumberInput, string][]).forEach(([key, value]) => {
-      if (!isNumber(value)) {
-        newIsValid[key] = false;
-        if (!hasError) {
-          message = '카드 번호는 숫자만 입력해주세요.';
-          hasError = true;
-        }
-      } else if (!isFourDigits(value)) {
-        newIsValid[key] = false;
-        if (!hasError) {
-          message = '카드 번호는 4자리로 입력해주세요.';
-          hasError = true;
-        }
-      } else if (key !== 'input1') {
-        newIsValid[key] = true;
-      }
-    });
-
-    setIsValid(newIsValid);
-    setErrorMessage(message);
+    validateCardNumber({ numbers, isValid, setIsValid, setErrorMessage });
   };
 
   return { cardNumber, setCardNumber, handleCardNumber, isValid, errorMessage };
-};
-
-const isNumber = (value: string) => !Number.isNaN(Number(value));
-
-const isFourDigits = (value: string) => value.length === 4;
-
-const isValidFirstCardNumber = (value: string) => {
-  if (!/^\d+$/.test(value)) return false;
-  if (value.startsWith('4')) return true;
-  const prefix = Number(value.slice(0, 2));
-  return prefix >= 51 && prefix <= 55;
 };
 
 export default useCardNumber;
